Initial work to support ESM
