+++ .babelrc
@@ -4,10 +4,14 @@
       "@babel/preset-env",
       {
         "targets": {
-          "node": "8"
-        }
+          "node": "14"
+        },
+        "modules": false
       }
     ]
   ],
-  "plugins": [["@babel/plugin-transform-runtime", { "corejs": 3 }]]
+  "plugins": [["@babel/plugin-transform-runtime", {
+    "corejs": 3,
+    "useESModules": true
+  }]]
 }
