+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.26.3",
   "description": "programmatic plopping for fun and profit",
   "main": "lib/index.js",
+  "type": "module",
   "types": "types/index.d.ts",
   "scripts": {
     "clean": "node ./build-scripts/clean",
@@ -39,18 +40,18 @@
   },
   "homepage": "https://github.com/amwmedia/node-plop#readme",
   "engines": {
-    "node": ">=8.9.4"
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.6.4",
-    "@babel/core": "^7.6.4",
-    "@babel/preset-env": "^7.6.3",
-    "@babel/register": "^7.6.2",
-    "@babel/plugin-transform-runtime": "^7.9.0",
+    "@babel/cli": "^7.16.0",
+    "@babel/core": "^7.16.0",
+    "@babel/preset-env": "^7.16.4",
+    "@babel/register": "^7.16.0",
+    "@babel/plugin-transform-runtime": "^7.16.4",
     "@types/node": "^13.13.5",
     "@typescript-eslint/eslint-plugin": "^2.21.0",
     "@typescript-eslint/parser": "^2.21.0",
-    "ava": "^2.4.0",
+    "ava": "^4.0.0-rc.1",
     "dtslint": "^3.5.1",
     "eslint": "^6.8.0",
     "eslint-config-standard": "^14.1.0",
@@ -64,7 +65,7 @@
     "typescript": "^3.8.2"
   },
   "dependencies": {
-    "@babel/runtime-corejs3": "^7.9.2",
+    "@babel/runtime-corejs3": "^7.16.3",
     "@types/inquirer": "^6.5.0",
     "change-case": "^3.1.0",
     "del": "^5.1.0",
@@ -80,12 +81,9 @@
     "files": [
       "tests/*.ava.js"
     ],
-    "sources": [
-      "tests/**/plopfile.js",
-      "lib/**/*.js"
-    ],
-    "require": [
-      "@babel/register"
+    "ignoredByWatcher": [
+      "lib/**/*",
+      "node_modules/**/*"
     ],
     "tap": true
   }
