+++ src/actions/_common-action-add-file.js
@@ -8,7 +8,7 @@ import {
 	getRelativeToBasePath
 } from './_common-action-utils';
 import {isBinaryFileSync} from 'isbinaryfile';
-import * as fspp from '../fs-promise-proxy';
+import * as fspp from '../fs-promise-proxy.js';
 
 export default async function addFile(data, cfg, plop) {
 	const fileDestPath = makeDestPath(data, cfg, plop);
