+++ tests/add-action-executable-file.ava.js
@@ -1,5 +1,5 @@
 import fs from 'fs';
-import AvaTest from './_base-ava-test';
+import AvaTest from './_base-ava-test.js';
 const { test, mockPath, testSrcPath, nodePlop } = new AvaTest(__filename);
 
 const plop = nodePlop();
@@ -38,7 +38,7 @@ if (process.platform !== 'win32') {
 			});
 
 			await plop.getGenerator('addExecutable').runActions();
-			
+
 			const destStats = fs.statSync(`${testSrcPath}/added.sh`);
 			t.is(destStats.mode & fs.constants.S_IXUSR, fs.constants.S_IXUSR);
 		}
