+++ tests/add-action-no-template.ava.js
@@ -1,6 +1,6 @@
-import * as fspp from '../src/fs-promise-proxy';
+import * as fspp from '../src/fs-promise-proxy.js';
 import path from 'path';
-import AvaTest from './_base-ava-test';
+import AvaTest from './_base-ava-test.js';
 const {test, testSrcPath, nodePlop} = (new AvaTest(__filename));
 
 const plop = nodePlop();
