+++ tests/custom-data-in-actions.ava.js
@@ -1,7 +1,7 @@
 // import fs from 'fs';
-import * as fspp from '../src/fs-promise-proxy';
+import * as fspp from '../src/fs-promise-proxy.js';
 import path from 'path';
-import AvaTest from './_base-ava-test';
+import AvaTest from './_base-ava-test.js';
 const {test, mockPath, testSrcPath, nodePlop} = (new AvaTest(__filename));
 
 const plop = nodePlop(`${mockPath}/plopfile.js`);
