+++ tests/force-del-outside-cwd.ava.js
@@ -1,5 +1,5 @@
 import fs from 'fs';
-import AvaTest from './_base-ava-test';
+import AvaTest from './_base-ava-test.js';
 const {test, mockPath, testSrcPath, nodePlop} = (new AvaTest(__filename));
 
 const plop = nodePlop(`${mockPath}/sub/plopfile.js`);
