+++ tests/prompt-bypass-list.ava.js
@@ -1,5 +1,5 @@
-import AvaTest from './_base-ava-test';
-import promptBypass from '../lib/prompt-bypass';
+import AvaTest from './_base-ava-test.js';
+import promptBypass from '../lib/prompt-bypass.js';
 
 const {test, nodePlop} = (new AvaTest(__filename));
 const plop = nodePlop();
@@ -63,7 +63,7 @@ test('verify good bypass input', function (t) {
 });
 
 test('verify bad bypass input', function (t) {
-	t.throws(() => promptBypass(prompts, ['asdf'], plop));
-	t.throws(() => promptBypass(prompts, ['6'], plop));
-	t.throws(() => promptBypass(prompts, [6], plop));
+	t.throws(() => promptBypass(prompts, ['asdf'], {is: plop}));
+	t.throws(() => promptBypass(prompts, ['6'], {is: plop}));
+	t.throws(() => promptBypass(prompts, [6], {is: plop}));
 });
